requirements.txt
pytest
pytest-playwright
allure-pytest
pytest-base-url
python-slugify
playwright-stealth
faker
pytest-xdist